Getting an Std from Javascript
Copying javascript's std folder is a dirty way of bootstrapping standard
classes and types, but some of this code can be reused as-is.  It'll be
easier to explain things in terms of what needs to change, rather than
what needs to be added.

I'm not adding all of the js-specific stuff dealing with browsers, html,
standard libraries like jquery, or utils like swfobject.