hello world, in Lua!
This commit makes the minimal changes necessary to create a "hello
world" program, compile it to lua, and execute it successfully.

```haxe
class Main {
    static function main() {
        trace("hello world");
    }
}
```

The good:

1) The lua target can (mostly) reuse javascript's object-based
namespacing and prototype based specification.
2) minor syntax differences (x ~= false instead of x != false, x .. ''
    instead of x + '', etc.)
3) dealing with minor syntax quirks ("end" keywords, missing brackets,
    etc) turns out to be not so bad.

The bad:
1) Classes/types don't seem to always be generated in the right order.
I'm getting lucky with the simple hello world example since there's not
much of an import hierarchy.
2) I get tripped up by Ocaml's strange syntax... constantly.

The ugly:

1) Lua doesn't have any sort of increment/decrement operators, nor does
it have += operators.  This turns out to be an incredible pain since
assignment is a statement (and not an expression), and there doesn't
seem to be any way to mutate a local unless explicitly specified in lua
script.  Solving this is easily the biggest problem I see.  Imagine an
expression like:

```
var x = 1;
trace(x++ + foo(x) + ++x);
```

The first part of the expression evaluates to 1, but the value "x" is
incremented immediately afterwards, and passed to "foo".  Then x is
concatenated with another prefix-incremeted version of itself.

In short, it's not possible to make these kinds of changes to x
before/after the given expression.  The increment operations need to be
done at the time of evaluation *within* the expression.

The solution I can see involves generate inline increment handlers for
each variable that is affected by a unary inc/dec operator and/or
OpAssign. E.g., the following is valid lua code, and seems to be the
only way of implementing this expression correctly.

```
var x = 1;
function op_x(value, pre)
    if pre then
       x = value
       return x
    else
       local oldx = x
       x = value
       return oldx
    end
end
trace(op_x(x+1,false) + foo(x) + op_x(x+1, true));
```

Unfortunately, that's a lot of boilerplate for such a simple operator.
It's going to make reading generated lua code that contains these
expressions very difficult.  Also, the compiler will have to read ahead
throuh an entire expression to see if it needs to generate one or more
of these operator assignment functions instead of the raw value.

There's more "ugly", of course, but it all pales in comparison to this.
Hopefully there's a better solution.