Lua: Fix String for new OOP mechanism
I tried a new approach for OOP in a previous commit.  It works ok, but
broke strings.  It forced me to reconsider how to set up the String
class for the Haxe Lua target.

The way things work now is that the base lua String.hx class sets the
metatable for the lua "string" class to be the String.hx metatable.
I.e., field accesses will be referenced against the base lua string
instance first, and then against the Haxe String metatable methods.

The Haxe String metatable methods also include an __index method which
looks for references to a field "length", and aliases this against the
respective lua implementation.