Lua : Misc wip on getting reflection/typing to work.
There's still a lot of work to do on getting reflection to work 100% on
Lua.  This changelist adds a few utility functions, and cleans up some
code in genlua, so at least it's a step in the right direction.

The main concern now is that lua has no mechanism for specifying a
null-valued field in the run time.  That is, if the field is null, you
can't tell if the field exists at all.

One approach is just to index all of the fields at compile time, but
this adds a good deal of code, and doesn't cover all of the other ways
in which one needs to access runtime field information.