HttpJs handle http request with r.status == 0 (#8759)
* HttpJs handle http request with r.status == 0

or `r.response == null`.
this will fix #8749 
Additional it will handle any case when `r.response == null`. The later would fix the case for a http request which could not be resolved too, But handling the particular case when `r.status == 0` will give the right error message.

* rev HttpJs handle http request with r.status == 0 

or `r.response == null`.
this will fix HaxeFoundation#8749 
Additional it will handle any case when `r.response == null`. The later would fix the case for a http request which could not be resolved too, But handling the particular case when `r.status == 0` will give the right error message.